refactor: handle key press as keyboard util
