+++ src/script/util/KeyboardUtil.ts
@@ -118,3 +118,54 @@ export const handleEnterDown = (event: React.KeyboardEvent<HTMLElement> | Keyboa
   }
   return true;
 };
+
+/**
+ * Handles global key press event - calls onPress when the key is pressed and onRelease when the key is released.
+ * @note The onPress is called only once when the key is pressed (it is not called repeatedly when key is held down).
+ * @param key The key to listen to.
+ * @param onPress A function to call when the key is pressed.
+ * @param onRelease A function to call when the key is released.
+ * @returns A function to unsubscribe.
+ */
+export const handleKeyPress = (key: string, {onPress, onRelease}: {onPress: () => void; onRelease: () => void}) => {
+  let isKeyDown = false;
+
+  const handleKeyDown = (event: KeyboardEvent) => {
+    if (event.key !== key) {
+      return;
+    }
+
+    // Do nothing if the key press was already registered.
+    if (isKeyDown) {
+      return;
+    }
+
+    isKeyDown = true;
+
+    onPress();
+  };
+
+  const handleKeyUp = (event: KeyboardEvent) => {
+    if (event.key !== key) {
+      return;
+    }
+
+    // If the key was not pressed, we do nothing.
+    if (!isKeyDown) {
+      return;
+    }
+
+    // Release the key.
+    isKeyDown = false;
+
+    onRelease();
+  };
+
+  window.addEventListener('keydown', handleKeyDown);
+  window.addEventListener('keyup', handleKeyUp);
+
+  return () => {
+    window.removeEventListener('keydown', handleKeyDown);
+    window.removeEventListener('keyup', handleKeyUp);
+  };
+};
